fix: move fixtures to `unit/conftest.py` to avoid overriding and call order issues
